--- @stellar/lyra-api/src/index.ts
@@ -1,5 +1,0 @@
-import { getPublicKey } from "./getPublicKey";
-import { signTransaction } from "./signTransaction";
-import { isConnected } from "./isConnected";
-
-export { getPublicKey, signTransaction, isConnected };
