--- @stellar/lyra-api/src/isConnected.ts
@@ -1,7 +1,0 @@
-declare global {
-  interface Window {
-    lyra: boolean;
-  }
-}
-
-export const isConnected = () => window.lyra;
