--- @stellar/lyra-api/src/signTransaction.ts
@@ -1,14 +1,0 @@
-import { submitTransaction } from "@lyra/api/external";
-import { ExternalRequest } from "@lyra/api/types";
-
-export const signTransaction = async (params: ExternalRequest) => {
-  let response = { transactionStatus: "", error: "" };
-
-  try {
-    response = await submitTransaction(params);
-  } catch (e) {
-    console.error(e);
-  }
-
-  return response;
-};
