--- @stellar/lyra-api/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "sourceRoot": "./src",
-    "baseUrl": "./src",
-    "declaration": true,
-    "declarationDir": "build/",
-    "outDir": "build/"
-  },
-  "include": ["src"],
-  "exclude": ["build", "node_modules"]
-}
