+++ extension/src/contentScript/lyraApi/status.ts
@@ -1,0 +1,3 @@
+export const status = () => {
+  // this method still to come
+};
